project/backend/requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_cors==4.0.0
psycopg2-binary==2.9.9
flask-socketio==5.3.6
eventlet==0.35.2
python-socketio==5.10.0
python-dotenv==1.0.0